beards/digocean/requirements.txt
python-digitalocean==1.10.1
